refactor: member 패키지 내 파일들 패키지화
